+++ code/app/src/test/java/com/hamidat/nullpointersapp/NotificationFragmentTest.java
@@ -1,7 +1,6 @@
 package com.hamidat.nullpointersapp;
 
 import static org.junit.Assert.*;
-import com.hamidat.nullpointersapp.utils.testUtils.TestHelper;
 
 import org.junit.Test;
